compass-runtime/requirements.txt
azure-identity==1.25.1
azure-storage-blob==12.27.1
azure-storage-queue==12.14.1
bidict==0.23.1
cryptography==46.0.3
cudf-cu12==25.12.0
duckdb==1.4.3
filelock==3.20.1
finai==0.0.1 # rc:ignore
fsspec==2025.12.0
hydra-core==1.3.2
jinja2==3.1.6
markupsafe==3.0.3
matplotlib==3.10.8
mlflow==3.5.0 # rc:ignore
mpmath==1.3.0
networkx==3.6.1
numpy==2.2.6 # rc:ignore numba 0.61.2 depends on numpy<2.3 and >=1.24
nvidia-ml-py==13.590.44
omegaconf==2.3.0
openpyxl==3.1.5
pandas==2.3.3
polars==1.36.1
pyaml==25.7.0
pyarrow==21.0.0 # rc:ignore mlflow 3.5.0 depends on pyarrow<22 and >=4.0.0
PyJWT==2.10.1
python-dateutil==2.9.0.post0
pytz==2025.2
pyyaml==6.0.3
rabbitmq_amqp_python_client==0.4.0
requests==2.32.5
scipy==1.16.3
setuptools==80.9.0
six==1.17.0
sympy==1.13.1 # rc:ignore torch 2.6.0 depends on sympy==1.13.1
tensorboard==2.20.0
torch==2.6.0 # rc:ignore
triton==3.2.0  # rc:ignore torch 2.6.0 requires triton==3.2.0
typing_extensions==4.15.0
tzdata==2025.3
